Add stubs for functions in stat.c
In case stat.o is optionally not compiled, stub functions defined in
fs.h will make sure the rest of the kernel will still compile. These
stub functions return permission errors, as they are unimplemented to
reduce kernel size.

If stat.o is setup to be compiled, then fs.h will define the appropriate
function prototypes.

This method will make sure that the tinification efforts have limited
impact of the maintainability of the subsystems to be compiled out.